move csharp demo to demo folder
