+++ csrc/apis/csharp/MMDeploySharp.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 16.0.31729.503
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MMDeploySharp", "src\MMDeploySharp\MMDeploySharp.csproj", "{3DC914EB-A8FB-4A89-A7CF-7DF9CC5284A6}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Demo", "src\Demo\Demo.csproj", "{661A4295-68CF-41C6-85B8-B11748113A6F}"
-EndProject
 
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
