--- csrc/apis/csharp/src/Demo/Demo.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net48</TargetFramework>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="OpenCvSharp4.runtime.win" Version="4.5.5.20211231" />
-    <PackageReference Include="OpenCvSharp4" Version="4.5.5.20211231" />
-    <ProjectReference Include="..\MMDeploySharp\MMDeploySharp.csproj" PrivateAssets="All" />
-  </ItemGroup>
-
-</Project>
